pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monarch-mcp-server"
version = "0.1.0"
description = "Model Context Protocol (MCP) server for Monarch Money personal finance platform"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Rob Cerda" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0.0",
    "monarchmoneycommunity>=1.3.0",
    "keyring>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/robcerda/monarch-mcp-server"
Repository = "https://github.com/robcerda/monarch-mcp-server"
Issues = "https://github.com/robcerda/monarch-mcp-server/issues"

[project.scripts]
monarch-mcp-server = "monarch_mcp_server.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
